All size_t in bitsize changed to bufflen_t type

+++ tests/BitStreamTests.cpp
@@ -60,7 +60,7 @@ TEST(BitStream, typesSerialization)
 
 TEST(BitStream, BoolTest)
 {
-    typedef uint64_t addType;
+    typedef uint8_t addType;
     addType num = 234;
     addType newNum;
